adstract functions to reuse code
